backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
supabase==2.11.0
sentence-transformers>=2.2
xgboost>=2.0
pandas>=2.1
numpy>=1.26
playwright>=1.40
scrapy>=2.11
google-generativeai>=0.3
python-decouple>=3.8
gunicorn>=21.2
pdfminer.six>=20221105
scikit-learn>=1.3
joblib>=1.3
python-dotenv>=1.0
